backend/requirements.txt
streamlit
requests
python-dotenv
fastapi
uvicorn

